+++ inc/product.php
@@ -7,7 +7,7 @@ $total_records_per_page = 20;
 
 error_reporting(E_ALL ^ E_NOTICE); 
 $product = 'JBs Perfumers Vault';
-$ver = '0.1.1 Beta';
+$ver = '0.1.2 Beta';
 
 $top_n = '25';
 $heart_n = '50';