+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-16/05/2020 - Checking of critical ingredient data missing
+v0.1.2 - Add: CSV import function
+		 FIX: metadata delete
+
+v0.1.1 - ADD: Checking of critical ingredient data missing
+
 15/05/2020 - Pyramid view added
 13/05/2020 - upload and tmp directory structure added
 12/05/2020 - 0.1 Beta public release